+++ src/main/java/com/codeborne/selenide/Command.java
@@ -6,6 +6,6 @@ import java.io.IOException;
 
 public interface Command<T> {
   Object[] NO_ARGS = new Object[0];
-  
+
   T execute(SelenideElement proxy, WebElementSource locator, Object[] args) throws IOException;
 }